+++ include/optimizer.h
@@ -5,11 +5,11 @@
 
 class Optimizer {
  protected:
-  float lr;  // learning rate
-  float decay;  // weight decay factor (default: 0)
+  double lr;  // learning rate
+  double decay;  // weight decay factor (default: 0)
 
  public:
-  explicit Optimizer(float lr = 0.01, float decay = 0.0) :
+  explicit Optimizer(double lr = 0.01, double decay = 0.0) :
                      lr(lr), decay(decay) {}
   virtual ~Optimizer() {}
