+++ src/cross_entropy_los.cpp
@@ -2,7 +2,7 @@
 
 void CrossEntropy::evaluate(const Matrix& pred, const Matrix& target) {
   int n = pred.cols();
-  const float eps = 1e-8;
+  const double eps = 1e-8;
   // forward: L = \sum{ -y_i*log(p_i) } / n
   loss = - (target.array().cwiseProduct((pred.array() + eps).log())).sum();
   loss /= n;
